model/requirements.txt
torch>=2.1.0
torchvision>=0.16.0
Pillow>=10.2.0
numpy>=1.26.0
matplotlib>=3.8.0
tqdm>=4.66.0
